Add State & Haptics throughout App
Currently, the app was just a shell without any State. We can track
state by creating the Shop ObservableObject and injecting it as an
EnvironmentObject throughout our app.  We have access to whether
a product was tapped, the specific selected Product and tracking
quantity of products added to the cart.